requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytest-django==4.5.2
pytest==7.2.2
pytest-cov==4.1.0
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7

